requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
certifi==2024.2.2
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Django==5.0
google==3.0.0
googletrans==3.0.0
gTTS==2.5.1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.2.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
mysql-connector==2.2.9
mysql-connector-python==8.3.0
package_name==0.1
protobuf==4.21.12
pywin32==306
requests==2.31.0
rfc3986==1.5.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.4
termcolor==2.4.0
tzdata==2023.3
urllib3==2.2.1
